+++ Libraries/drivers/K/inc/ftm.h
@@ -120,10 +120,10 @@ typedef struct
 
 
 //!< API functions
-void FTM_PWM_ChangeDuty(uint8_t instance, uint8_t chl, uint32_t pwmDuty);
+void FTM_PWM_ChangeDuty(uint32_t instance, uint8_t chl, uint32_t pwmDuty);
 uint8_t FTM_PWM_QuickInit(uint32_t FTMxMAP, uint32_t frequencyInHZ);
 void FTM_PWM_Init(FTM_PWM_InitTypeDef* FTM_InitStruct);
-void FTM_PWM_InvertPolarity(uint8_t instance, uint8_t chl, uint32_t config);
+void FTM_PWM_InvertPolarity(uint32_t instance, uint8_t chl, uint32_t config);
 uint32_t FTM_QD_QuickInit(uint32_t FTMxMAP);
 void FTM_QD_Init(FTM_QD_InitTypeDef * FTM_QD_InitStruct);
 void FTM_QD_GetData(uint32_t instance, uint32_t* value, uint8_t* direction);
